+++ data/claudia.txt
@@ -1,1 +1,5 @@
-T | 0 | oeooe
+T | 1 | read book
+D | 0 | return book | 2024-01-12 18:00
+E | 0 | project meeting | 2024-01-18 18:00 | 2024-01-20 18:00
+T | 0 | join sports club
+T | 0 | borrow book
