+++ src/main/java/claudia/command/UnmarkCommand.java
@@ -1,0 +1,41 @@
+package claudia.command;
+
+import claudia.exception.ClaudiaException;
+import claudia.exception.InvalidFormatException;
+import claudia.exception.InvalidTaskNumberException;
+import claudia.misc.TaskList;
+import claudia.storage.Storage;
+import claudia.task.Task;
+import claudia.ui.Ui;
+
+public class UnmarkCommand extends Command {
+    private final String index;
+
+    public UnmarkCommand(String index) {
+        this.index = index;
+    }
+
+    @Override
+    public TaskList execute(TaskList tasks, Ui ui, Storage storage) throws ClaudiaException {
+        try {
+            int i = Integer.parseInt(index) - 1; // zero-based
+            if (i < 0 || i >= tasks.size()) {
+                throw new InvalidTaskNumberException(tasks.size());
+            }
+
+            Task t = tasks.getTask(i);
+            t.markAsNotDone();
+            storage.save(tasks);
+            ui.showUnmark(t);
+        } catch (NumberFormatException e) {
+            throw new InvalidFormatException("Invalid number.");
+        }
+
+        return tasks;
+    }
+
+    @Override
+    public boolean isExit() {
+        return false;
+    }
+}
