+++ src/main/java/claudia/exception/EmptyListException.java
@@ -1,0 +1,7 @@
+package claudia.exception;
+
+public class EmptyListException extends ClaudiaException {
+    public EmptyListException() {
+        super("The list is empty~ Add a task now!");
+    }
+}
