+++ src/main/java/claudia/exception/InvalidFormatException.java
@@ -1,0 +1,11 @@
+package claudia.exception;
+
+public class InvalidFormatException extends ClaudiaException {
+    public InvalidFormatException(String message) {
+        super(message);
+    }
+}
+
+// mark or unmark invalid int type
+// deadline missing /by
+// event missing /from and/or /to
