+++ src/main/java/claudia/exception/InvalidTaskNumberException.java
@@ -1,0 +1,7 @@
+package claudia.exception;
+
+public class InvalidTaskNumberException extends ClaudiaException {
+    public InvalidTaskNumberException(int noOfTasks) {
+        super(String.format("Enter a valid number between 1 to %d", noOfTasks));
+    }
+}
