+++ src/main/java/claudia/exception/MissingDescriptionException.java
@@ -1,0 +1,7 @@
+package claudia.exception;
+
+public class MissingDescriptionException extends ClaudiaException {
+    public MissingDescriptionException(String command) {
+        super(String.format("The description of a %s cannot be empty.", command));
+    }
+}
