--- src/main/java/exceptions/ClaudiaException.java
@@ -1,11 +1,0 @@
-package exceptions;
-
-public class ClaudiaException extends Exception {
-    public ClaudiaException(String message) {
-        super(message);
-    }
-}
-
-// if it is an exception
-// extends Exception -> checked exception
-// declared in throws or caught in try-catch block
