--- src/main/java/exceptions/EmptyListException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class EmptyListException extends ClaudiaException {
-    public EmptyListException() {
-        super("The list is empty~ Add a task now!");
-    }
-}
