--- src/main/java/exceptions/InvalidFormatException.java
@@ -1,11 +1,0 @@
-package exceptions;
-
-public class InvalidFormatException extends ClaudiaException {
-    public InvalidFormatException(String message) {
-        super(message);
-    }
-}
-
-// mark or unmark invalid int type
-// deadline missing /by
-// event missing /from and/or /to
