--- src/main/java/exceptions/InvalidTaskNumberException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class InvalidTaskNumberException extends ClaudiaException {
-    public InvalidTaskNumberException(int noOfTasks) {
-        super(String.format("Enter a valid number between 1 to %d", noOfTasks));
-    }
-}
