--- src/main/java/exceptions/MissingDescriptionException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class MissingDescriptionException extends ClaudiaException {
-    public MissingDescriptionException(String command) {
-        super(String.format("The description of a %s cannot be empty.", command));
-    }
-}
