--- src/main/java/exceptions/UnknownInputException.java
@@ -1,10 +1,0 @@
-package exceptions;
-
-public class UnknownInputException extends ClaudiaException {
-    public UnknownInputException() {
-        super("I don't know what that means :(");
-    }
-}
-
-// empty input
-// command unknown -> use enums
