--- src/main/java/misc/TaskList.java
@@ -1,41 +1,0 @@
-package misc;
-
-import tasks.Task;
-
-import java.util.ArrayList;
-
-public class TaskList {
-    private ArrayList<Task> tasks;
-
-    public TaskList() {
-        this.tasks = new ArrayList<>();
-    }
-
-    public TaskList(ArrayList<Task> tasks) {
-        this.tasks = tasks;
-    }
-
-    public void addTask(Task task) {
-        tasks.add(task);
-    }
-
-    public void removeTask(int index) {
-        tasks.remove(index);
-    }
-
-    public ArrayList<Task> getTasks() {
-        return tasks;
-    }
-
-    public int size() {
-        return tasks.size();
-    }
-
-    public boolean isEmpty() {
-        return tasks.isEmpty();
-    }
-
-    public Task getTask(int index) {
-        return tasks.get(index);
-    }
-}
